Moved functions outside of the main file and made getRatio work
