+++ functions/gameModeStatsDict.py
@@ -1,0 +1,5 @@
+gamemodeStats = {
+    "SkyWars": ['kills', 'deaths', 'wins'],
+    "Bedwars": ['kills_bedwars', 'deaths_bedwars', 'wins_bedwars'],
+    "HungerGames": ['kills', 'deaths', 'wins']
+}
