+++ functions/getGameMode.py
@@ -1,3 +1,4 @@
+
 def inputGameMode():
     return input("What game mode would you like to search for? (Skywars, Bedwars, HungerGames)")
